move call into codewhispererService
